1.42 Quick patch
Quick patch for compatibility with 1.42, due to not wanting to use Maven and relying on Apache Common File Utils I will be migrating the buttons from flat files to SQLite DBs, this update will break those existing buttons but future updates will not break the SQLite version. This plugin will function fine as is, but will not have Jungle buttons enabled until the next full release of MineSwarm (hopefully this week).